feat(ingest): update with suggestions, split up code into smaller functions with more docs
