+++ kubernetes/loculus/values.yaml
@@ -1568,7 +1568,7 @@ defaultOrganisms:
         taxon_id: 3052518
         nextclade_dataset_server: https://raw.githubusercontent.com/nextstrain/nextclade_data/cornelius-cchfv/data_output
         nextclade_dataset_name: nextstrain/cchfv/linked
-        grouping_ground_truth: https://anna-parker.github.io/influenza-a-groupings/results/cchf_groups.json
+        grouping_override: https://anna-parker.github.io/influenza-a-groupings/results/cchf_groups.json
     enaDeposition:
       configFile:
         taxon_id: 3052518
